--- a/Common/Securities/Future/FuturesExpiryFunctions.cs
+++ b/Common/Securities/Future/FuturesExpiryFunctions.cs
@@ -1,11 +1,11 @@
 ﻿/*
  * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
  * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
- * 
- * Licensed under the Apache License, Version 2.0 (the "License"); 
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -29,9 +29,9 @@ namespace QuantConnect.Securities.Future
         /// </summary>
         public static Func<DateTime, DateTime> FuturesExpiryFunction(string symbol)
         {
-            if (FuturesExpiryDictionary.ContainsKey(symbol.ToUpper()))
+            if (FuturesExpiryDictionary.ContainsKey(symbol.LazyToUpper()))
             {
-                return FuturesExpiryDictionary[symbol.ToUpper()];
+                return FuturesExpiryDictionary[symbol.LazyToUpper()];
             }
 
             // If function for expiry cannot be found pass the date through
@@ -549,7 +549,7 @@ namespace QuantConnect.Securities.Future
             // MXN (6M): https://www.cmegroup.com/trading/fx/emerging-market/mexican-peso_contract_specifications.html
             {Futures.Currencies.MXN, (time =>
                 {
-                    // 9:16 a.m. Central Time (CT) on the second business day immediately preceding the third Wednesday of the contract month (usually Monday). 
+                    // 9:16 a.m. Central Time (CT) on the second business day immediately preceding the third Wednesday of the contract month (usually Monday).
                     var thirdWednesday = FuturesExpiryUtilityFunctions.ThirdWednesday(time);
                     var secondBusinessDayPrecedingThirdWednesday = FuturesExpiryUtilityFunctions.AddBusinessDays(thirdWednesday,-2);
 
@@ -703,7 +703,7 @@ namespace QuantConnect.Securities.Future
             // E-mini Euro FX (E7): https://www.cmegroup.com/trading/fx/g10/e-mini-euro-fx_contract_specifications.html
             {Futures.Currencies.EuroFXEmini, (time =>
                 {
-                    // 9:16 a.m. Central Time (CT) on the second business day immediately preceding the third Wednesday of the contract month (usually Monday). 
+                    // 9:16 a.m. Central Time (CT) on the second business day immediately preceding the third Wednesday of the contract month (usually Monday).
                     var thirdWednesday = FuturesExpiryUtilityFunctions.ThirdWednesday(time);
                     var secondBusinessDayPrecedingThirdWednesday = FuturesExpiryUtilityFunctions.AddBusinessDays(thirdWednesday, -2);
                     var holidays = MarketHoursDatabase.FromDataFolder()
@@ -722,7 +722,7 @@ namespace QuantConnect.Securities.Future
             // Euro/Australian Dollar (EAD): https://www.cmegroup.com/trading/fx/g10/euro-fx-australian-dollar_contract_specifications.html
             {Futures.Currencies.EURAUD, (time =>
                 {
-                    // 9:16 a.m. Central Time (CT) on the second business day immediately preceding the third Wednesday of the contract month (usually Monday). 
+                    // 9:16 a.m. Central Time (CT) on the second business day immediately preceding the third Wednesday of the contract month (usually Monday).
                     var thirdWednesday = FuturesExpiryUtilityFunctions.ThirdWednesday(time);
                     var secondBusinessDayPrecedingThirdWednesday = FuturesExpiryUtilityFunctions.AddBusinessDays(thirdWednesday, -2);
                     var holidays = MarketHoursDatabase.FromDataFolder()
@@ -741,7 +741,7 @@ namespace QuantConnect.Securities.Future
             // Euro/Canadian Dollar (ECD): https://www.cmegroup.com/trading/fx/g10/euro-fx-canadian-dollar_contract_specifications.html
             {Futures.Currencies.EURCAD, (time =>
                 {
-                    // Trading terminates at 9:16 a.m. CT on the second business day prior to the third Wednesday of the contract month. 
+                    // Trading terminates at 9:16 a.m. CT on the second business day prior to the third Wednesday of the contract month.
                     var thirdWednesday = FuturesExpiryUtilityFunctions.ThirdWednesday(time);
                     var secondBusinessDayPrecedingThirdWednesday = FuturesExpiryUtilityFunctions.AddBusinessDays(thirdWednesday, -2);
                     var holidays = MarketHoursDatabase.FromDataFolder()
@@ -893,7 +893,7 @@ namespace QuantConnect.Securities.Future
                     return lastBusinessDay;
                 })
             },
-            // Argus Propane Far East Index Futures (A7E): 
+            // Argus Propane Far East Index Futures (A7E):
             {Futures.Energies.ArgusPropaneFarEastIndex, (time =>
                 {
                     // Trading shall cease on the last business day of the contract month.
@@ -1037,7 +1037,7 @@ namespace QuantConnect.Securities.Future
             // Argus LLS vs. WTI (Argus) Trade Month Futures (AE5): https://www.cmegroup.com/trading/energy/crude-oil/argus-lls-vs-wti-argus-trade-month-swap-futures_contract_specifications.html
             {Futures.Energies.ArgusLLSvsWTIArgusTradeMonth, (time =>
                 {
-                    // Trading shall cease at the close of trading on the last business day that falls on or before the 25th calendar day of the month prior to the contract month. If the 25th calendar day is a weekend or holiday, trading shall cease on the first business day prior to the 25th calendar day. 
+                    // Trading shall cease at the close of trading on the last business day that falls on or before the 25th calendar day of the month prior to the contract month. If the 25th calendar day is a weekend or holiday, trading shall cease on the first business day prior to the 25th calendar day.
                     var previousMonth = time.AddMonths(-1);
                     var twentyFifthDay = new DateTime(previousMonth.Year, previousMonth.Month, 25);
                     var holidays = MarketHoursDatabase.FromDataFolder()
@@ -1085,7 +1085,7 @@ namespace QuantConnect.Securities.Future
                     return FuturesExpiryUtilityFunctions.NthLastBusinessDay(time, 1);
                 })
             },
-            // Los Angeles CARB Diesel (OPIS) vs. NY Harbor ULSD (AKL): https://www.cmegroup.com/trading/energy/refined-products/los-angeles-carbob-diesel-opis-spread-swap_contract_specifications.html 
+            // Los Angeles CARB Diesel (OPIS) vs. NY Harbor ULSD (AKL): https://www.cmegroup.com/trading/energy/refined-products/los-angeles-carbob-diesel-opis-spread-swap_contract_specifications.html
             {Futures.Energies.LosAngelesCARBDieselOPISvsNYHarborULSD, (time =>
                 {
                     // Trading shall cease on the last business day of the contract month
@@ -1141,8 +1141,8 @@ namespace QuantConnect.Securities.Future
             // Mars (Argus) vs. WTI Trade Month (AYV): https://www.cmegroup.com/trading/energy/crude-oil/mars-crude-oil-argus-vs-wti-trade-month-spread-swap-futures_contract_specifications.html
             {Futures.Energies.MarsArgusVsWTITradeMonth, (time =>
                 {
-                    // Trading shall cease at the close of trading on the last business day that falls on or before the 25th calendar day of the 
-                    // month prior to the contract month. If the 25th calendar day is a weekend or holiday, trading shall cease on the 
+                    // Trading shall cease at the close of trading on the last business day that falls on or before the 25th calendar day of the
+                    // month prior to the contract month. If the 25th calendar day is a weekend or holiday, trading shall cease on the
                     // first business day prior to the 25th calendar day.
                     var twentyFifthDayPriorMonth = new DateTime(time.Year, time.Month, 25).AddMonths(-1);
                     var holidays = MarketHoursDatabase.FromDataFolder()
@@ -1318,7 +1318,7 @@ namespace QuantConnect.Securities.Future
             // Clearbrook Bakken Sweet Crude Oil Monthly Index (Net Energy) (CSW): https://www.cmegroup.com/trading/energy/crude-oil/clearbrook-bakken-crude-oil-index-net-energy_contract_specifications.html
             {Futures.Energies.ClearbrookBakkenSweetCrudeOilMonthlyIndexNetEnergy, (time =>
                 {
-                    // Trading terminates one Canadian business day prior to the Notice of Shipments (NOS) date on the Enbridge Pipeline. The NOS date occurs on or about the 20th calendar day of the month, subject to confirmation by Enbridge Pipeline. The official schedule for the NOS dates will be made publicly available by Enbridge. 
+                    // Trading terminates one Canadian business day prior to the Notice of Shipments (NOS) date on the Enbridge Pipeline. The NOS date occurs on or about the 20th calendar day of the month, subject to confirmation by Enbridge Pipeline. The official schedule for the NOS dates will be made publicly available by Enbridge.
                     // This report is behind a portal that requires registration (privately). As such, we cannot access the notice of shipment dates, but we can keep track
                     // of the CME group's website in order to discover the NOS dates
                     // Publication dates are also erratic. We must maintain a separate list from MHDB in order to keep track of these days
@@ -1401,7 +1401,7 @@ namespace QuantConnect.Securities.Future
             // Japan C&amp;F Naphtha (Platts) BALMO (E6): https://www.cmegroup.com/trading/energy/refined-products/japan-naphtha-balmo-swap-futures_contract_specifications.html
             {Futures.Energies.JapanCnFNaphthaPlattsBALMO, (time =>
                 {
-                    // Trading shall cease on the last business day of the contract month. 
+                    // Trading shall cease on the last business day of the contract month.
                     return FuturesExpiryUtilityFunctions.NthLastBusinessDay(time, 1);
                 })
             },
@@ -1416,11 +1416,11 @@ namespace QuantConnect.Securities.Future
                         .GetEntry("usa", Futures.Energies.Ethanol, SecurityType.Future)
                         .ExchangeHours
                         .Holidays;
-                    
+
                     while (daysCounted < 3)
                     {
                         // The asset continues trading on days contained within `USHoliday.Dates`, but
-                        // the last trade date is affected by those holidays. We check for 
+                        // the last trade date is affected by those holidays. We check for
                         // both MHDB entries and holidays to get accurate business days
                         if (holidays.Contains(currentDay) || USHoliday.Dates.Contains(currentDay))
                         {
@@ -1500,7 +1500,7 @@ namespace QuantConnect.Securities.Future
             {Futures.Meats.FeederCattle, (time =>
                 {
                     /* Trading shall terminate on the last Thursday of the contract month, except:
-                     * 1. The November contract shall terminate on the Thursday    
+                     * 1. The November contract shall terminate on the Thursday
                      * prior to Thanksgiving Day, unless a holiday falls on
                      * that Thursday or on any of the four weekdays prior to
                      * that Thursday, in which case trading shall terminate on
@@ -1572,7 +1572,7 @@ namespace QuantConnect.Securities.Future
             // Cash-Settled Cheese (CSC): https://www.cmegroup.com/trading/agricultural/dairy/cheese_contract_specifications.html
             {Futures.Dairy.CashSettledCheese, (time =>
                 {
-                    // Trading shall terminate on the business day immediately preceding the release date for the USDA monthly weighted average price in the U.S. for cheese. LTD close is at 12:10 p.m. Central Time 
+                    // Trading shall terminate on the business day immediately preceding the release date for the USDA monthly weighted average price in the U.S. for cheese. LTD close is at 12:10 p.m. Central Time
                     DateTime publicationDate;
                     if (DairyReportDates.TryGetValue(time, out publicationDate))
                     {
@@ -1620,7 +1620,7 @@ namespace QuantConnect.Securities.Future
             // Dry Whey (DY): https://www.cmegroup.com/trading/agricultural/dairy/dry-whey_contract_specifications.html
             {Futures.Dairy.DryWhey, (time =>
                 {
-                    // Trading shall terminate on the business day immediately preceding the day on which the USDA announces the Dry Whey price for that contract month. (LTD 12:10 p.m.) 
+                    // Trading shall terminate on the business day immediately preceding the day on which the USDA announces the Dry Whey price for that contract month. (LTD 12:10 p.m.)
                     DateTime publicationDate;
                     if (DairyReportDates.TryGetValue(time, out publicationDate))
                     {