--- a/Common/Securities/SymbolProperties.cs
+++ b/Common/Securities/SymbolProperties.cs
@@ -90,7 +90,7 @@ namespace QuantConnect.Securities
         /// <returns>A default instance of the<see cref="SymbolProperties"/> class</returns>
         public static SymbolProperties GetDefault(string quoteCurrency)
         {
-            return new SymbolProperties("", quoteCurrency.ToUpper(), 1, 0.01m, 1);
+            return new SymbolProperties("", quoteCurrency.LazyToUpper(), 1, 0.01m, 1);
         }
     }
 }